perbaiki sedikit

+++ app/Http/Controllers/ProfileController.php
@@ -70,7 +70,7 @@ class ProfileController extends Controller
             $mail->Port       = 587;                   // TCP port to connect to
         
             //Recipients
-            $mail->setFrom(getenv('EMAIL'), "Don't Replay");
+            $mail->setFrom(getenv('EMAIL'), "Don't reply");
             $mail->addAddress($email); // Add a recipient
 
             // Content
@@ -112,7 +112,7 @@ class ProfileController extends Controller
             $mail->Port       = 587;                   // TCP port to connect to
         
             //Recipients
-            $mail->setFrom(getenv('EMAIL'), "Don't Replay");
+            $mail->setFrom(getenv('EMAIL'), "Don't reply");
             $mail->addAddress($email); // Add a recipient
         
             // Content
@@ -155,7 +155,7 @@ class ProfileController extends Controller
             $mail->Port       = 587;                   // TCP port to connect to
         
             //Recipients
-            $mail->setFrom(getenv('EMAIL'), "Don't Replay");
+            $mail->setFrom(getenv('EMAIL'), "Don't reply");
             $mail->addAddress($email); // Add a recipient
         
             // Content
@@ -273,11 +273,11 @@ class ProfileController extends Controller
         $cleanedFileName = Str::slug($filename) . '_' . time() . '.' . $extension;
     
         $destinationPath = 'profileImage/';
-        $image_path = $destinationPath . $cleanedFileName;
+        $image_path = asset($destinationPath . $cleanedFileName);
     
         // Hapus file lama jika ada
         $oldImagePath = $user->image_path;
-        if ($oldImagePath && File::exists(public_path($oldImagePath))) {
+        if ($oldImagePath && File::exists(public_path($oldImagePath)) && $oldImagePath != '/profileImage/default.png') {
             File::delete(public_path($oldImagePath));
         }
